refactored test folder structure
